moves deskHash setting to local_storage middleware
