+++ assets/src/components/app/options/form/container.js
@@ -10,9 +10,10 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-	setPodId  : podId => { dispatch({ type: "SET_POD_ID", podId }) },
-	setDesk   : desk => { dispatch({ type: "SET_DESK", desk }) },
-	setCityId : cityId => { dispatch({ type: "SET_CITY_ID", cityId }) }
+	setPodId    : podId => { dispatch({ type: "SET_POD_ID", podId }) },
+	setDesk     : desk => { dispatch({ type: "SET_DESK", desk }) },
+	setCityId   : cityId => { dispatch({ type: "SET_CITY_ID", cityId }) },
+	setPassword : password => { dispatch({ type: "SET_PASSWORD", password }) }
 });
 
 export default connect(
